requirements.txt
# RAG core
fastembed
faiss-cpu
rank-bm25

# LangChain & LLM
langchain
langchain-community
groq

# Web/UI
streamlit

# Document processing
PyPDF2
python-docx

# Utilities
numpy
pandas
python-dotenv
requests